requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-generativeai==0.3.1
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
toncenter==0.0.1 